Data management

+++ Assets/Scripts/SlotsEngine/WinlineDefinition.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Data", menuName = "ScriptableObjects/WinlineDefinition")]
 public class WinlineDefinition : ScriptableObject
 {
-	[SerializeField] private float winMultiplier = 1;
+	[SerializeField] private int winMultiplier = 1;
 	[SerializeField] private int[] symbolIndexes;
 
-	public float WinMultiplier => winMultiplier;
+	public int WinMultiplier => winMultiplier;
 	public int[] SymbolIndexes => symbolIndexes;
 	public int Depth => symbolIndexes.Length;
 }
